data: refactor 3/7 protocols
